src/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
PyJWT==2.8.0
werkzeug==3.0.1
python-dotenv==1.0.0
pymysql==1.1.0
bcrypt==4.1.2  # 密码加密
cryptography==42.0.5